update appstrap and appserver to get up and running fix generator to use import instead of require
