+++ src/Appstrap.js
@@ -1,30 +1,13 @@
-const configLoader = require('./config/loader')
-const AppServer = require('./AppServer')
-const getPort = require('get-port')
+import configLoader from './config/loader'
+import AppServer from './AppServer'
 
 class Appstrap {
   constructor ({configPath, port = 5000, config = configLoader.load(configPath)}) {
-    this.port = port
-    this.config = config
-    this.server = new AppServer()
-    this.server.loadEndpoints(this.config.endpoints)
+    AppServer.configure({port, isSPA: !!config.bundle})
   }
 
-  async start () {
-    this.port = await getPort({port: this.port})
-    await this.server.httpServer.listenAsync(this.port)
-    console.log(`
-    ===============================================================
-      Appstrap loaded successfully.
-      A server has been started for you at the following address: 
-      http://localhost:${this.port}
-    ===============================================================
-    `)
-  }
-
-  async stop () {
-    await this.server.httpServer.closeAsync()
-  }
+  get start () { return AppServer.start }
+  get stop () { return AppServer.stop }
 }
 
-module.exports = Appstrap
+export default Appstrap
